+++ src/main/java/io/github/pmckeown/dependencytrack/Response.java
@@ -1,4 +1,4 @@
-package io.github.pmckeown.rest.model;
+package io.github.pmckeown.dependencytrack;
 
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
